Split gallery and reviews into components
Move the gallery and reviews sections into their own Angular components (ab-gallery-section and ab-reviews-section). Added new HTML/SCSS/TS files for both components, moved the gallery data/build logic and review data into their respective component classes, and relocated the related styles from the global _landing.scss into component-scoped SCSS. Updated the landing page to import and render the new components and removed the inlined templates/data and global styles, improving modularity and scoping of styles.
