+++ libs/adrian-badilla/frontend/landing-page/src/lib/adrian-badilla-ui-landing-page/adrian-badilla-ui-landing-page.component.html
@@ -340,30 +340,7 @@
   </div>
 
   <!-- ============ GALLERY (pinned horizontal scene) ============ -->
-  <section id="gallery" data-h-gallery style="position:relative;">
-    <div data-h-sticky style="position:sticky; top:0; height:100vh; overflow:hidden; display:flex; flex-direction:column; justify-content:center; gap:clamp(22px,4vh,46px); padding-block:clamp(40px,7vh,90px);">
-      <div style="max-width:1320px; width:100%; margin:0 auto; padding:0 clamp(20px,5vw,48px);">
-        <div style="display:flex; align-items:baseline; gap:16px; margin-bottom:14px;">
-          <span class="ab-chrome" style="font-family:'Anton',sans-serif; font-size:1.4rem;">03</span>
-          <span style="font-size:13px; font-weight:700; letter-spacing:.22em; text-transform:uppercase; color:var(--accent, hsl(356 88% 53%));">Resultados reales</span>
-        </div>
-        <h2 style="font-family:'Anton',sans-serif; font-weight:400; font-size:clamp(2.4rem,6vw,5rem); line-height:.92; letter-spacing:-.01em; text-transform:uppercase; margin:0;">Lográ tu objetivo</h2>
-        <p style="margin:14px 0 0; font-size:1rem; color:rgba(255,255,255,.55);">Seguí scrolleando — cada transformación entra en escena.</p>
-      </div>
-      <div style="overflow:hidden; width:100%; height:clamp(440px,88vh,960px); display:flex; align-items:center; -webkit-mask-image:linear-gradient(90deg, transparent, #000 9%, #000 91%, transparent); mask-image:linear-gradient(90deg, transparent, #000 9%, #000 91%, transparent);">
-        <div data-h-track style="display:flex; align-items:center; gap:clamp(56px,6vw,110px); width:max-content; padding:0 clamp(40px,12vw,200px); will-change:transform;">
-          @for (it of galleryItems; track $index) {
-            <figure class="g-item" [attr.data-depth]="it.depth" [attr.data-vy]="it.vy" [attr.data-align]="it.align" [style.width]="it.w" style="flex:none; margin:0;">
-              <div class="g-cap" style="font-size:11px; font-weight:700; letter-spacing:.2em; text-transform:uppercase; color:rgba(255,255,255,.5); margin-bottom:14px; white-space:nowrap;">{{ it.label }}</div>
-              <div class="g-frame" [style.height]="it.h" style="width:100%; border-radius:4px; overflow:hidden; position:relative; background:#0d0d0d;">
-                <img [src]="it.src" [alt]="it.label" loading="lazy" style="width:100%; height:100%; object-fit:cover; object-position:center; display:block;">
-              </div>
-            </figure>
-          }
-        </div>
-      </div>
-    </div>
-  </section>
+  <ab-gallery-section />
 
   <div data-scrub data-mask="left" data-srange="0,0.4" style="display:flex; align-items:center; gap:18px; max-width:1280px; margin:clamp(12px,5vh,48px) auto 0; padding:0 clamp(20px,5vw,48px);">
     <div style="flex:1; height:1px; background:linear-gradient(90deg, transparent, hsl(356 88% 53% / .4));"></div>
@@ -372,89 +349,7 @@
   </div>
 
   <!-- ============ REVIEWS ============ -->
-  <section id="reviews" style="margin:0 auto; padding:clamp(56px,9vh,120px) 0;">
-    <div data-scrub style="max-width:1280px; margin:0 auto clamp(40px,6vw,60px); padding:0 clamp(20px,5vw,48px);">
-      <div style="display:flex; align-items:baseline; gap:16px; margin-bottom:16px;">
-        <span class="ab-chrome" style="font-family:'Anton',sans-serif; font-size:1.4rem;">04</span>
-        <span style="font-size:13px; font-weight:700; letter-spacing:.22em; text-transform:uppercase; color:var(--accent, hsl(356 88% 53%));">5.0 ★ en Google</span>
-      </div>
-      <h2 data-scrub data-mask="up" style="font-family:'Anton',sans-serif; font-weight:400; font-size:clamp(2.4rem,5.5vw,4.6rem); line-height:.92; letter-spacing:-.01em; text-transform:uppercase; margin:0; max-width:18ch;">Historias que hablan solas</h2>
-    </div>
-
-    <div class="rev-row" style="overflow:hidden; -webkit-mask-image:linear-gradient(90deg, transparent, #000 5%, #000 95%, transparent); mask-image:linear-gradient(90deg, transparent, #000 5%, #000 95%, transparent);">
-      <div class="rev-track" style="display:flex; gap:20px; width:max-content; padding:14px 10px; animation:ab-marqueeR 56s linear infinite;">
-        @for (rev of reviews; track $index) {
-          <article class="rev-card">
-            <div class="rev-card__top">
-              <span style="color:var(--accent, hsl(356 88% 53%)); font-size:15px; letter-spacing:.14em;">★★★★★</span>
-              <span class="rev-g">G</span>
-            </div>
-            <p class="rev-quote">{{ rev.quote }}</p>
-            <div class="rev-foot">
-              <div class="rev-avatar">{{ rev.initials }}</div>
-              <div>
-                <div style="font-weight:700; font-size:.96rem; color:#fff;">{{ rev.name }}</div>
-                <div style="font-size:.82rem; color:rgba(255,255,255,.5);">{{ rev.role }}</div>
-              </div>
-            </div>
-          </article>
-        }
-        @for (rev of reviews; track $index) {
-          <article class="rev-card" aria-hidden="true">
-            <div class="rev-card__top">
-              <span style="color:var(--accent, hsl(356 88% 53%)); font-size:15px; letter-spacing:.14em;">★★★★★</span>
-              <span class="rev-g">G</span>
-            </div>
-            <p class="rev-quote">{{ rev.quote }}</p>
-            <div class="rev-foot">
-              <div class="rev-avatar">{{ rev.initials }}</div>
-              <div>
-                <div style="font-weight:700; font-size:.96rem; color:#fff;">{{ rev.name }}</div>
-                <div style="font-size:.82rem; color:rgba(255,255,255,.5);">{{ rev.role }}</div>
-              </div>
-            </div>
-          </article>
-        }
-      </div>
-    </div>
-
-    <div class="rev-row" style="overflow:hidden; margin-top:20px; -webkit-mask-image:linear-gradient(90deg, transparent, #000 5%, #000 95%, transparent); mask-image:linear-gradient(90deg, transparent, #000 5%, #000 95%, transparent);">
-      <div class="rev-track" style="display:flex; gap:20px; width:max-content; padding:14px 10px; animation:ab-marqueeR 72s linear infinite;">
-        @for (rev of reviewsB; track $index) {
-          <article class="rev-card">
-            <div class="rev-card__top">
-              <span style="color:var(--accent, hsl(356 88% 53%)); font-size:15px; letter-spacing:.14em;">★★★★★</span>
-              <span class="rev-g">G</span>
-            </div>
-            <p class="rev-quote">{{ rev.quote }}</p>
-            <div class="rev-foot">
-              <div class="rev-avatar">{{ rev.initials }}</div>
-              <div>
-                <div style="font-weight:700; font-size:.96rem; color:#fff;">{{ rev.name }}</div>
-                <div style="font-size:.82rem; color:rgba(255,255,255,.5);">{{ rev.role }}</div>
-              </div>
-            </div>
-          </article>
-        }
-        @for (rev of reviewsB; track $index) {
-          <article class="rev-card" aria-hidden="true">
-            <div class="rev-card__top">
-              <span style="color:var(--accent, hsl(356 88% 53%)); font-size:15px; letter-spacing:.14em;">★★★★★</span>
-              <span class="rev-g">G</span>
-            </div>
-            <p class="rev-quote">{{ rev.quote }}</p>
-            <div class="rev-foot">
-              <div class="rev-avatar">{{ rev.initials }}</div>
-              <div>
-                <div style="font-weight:700; font-size:.96rem; color:#fff;">{{ rev.name }}</div>
-                <div style="font-size:.82rem; color:rgba(255,255,255,.5);">{{ rev.role }}</div>
-              </div>
-            </div>
-          </article>
-        }
-      </div>
-    </div>
-  </section>
+  <ab-reviews-section />
 
   <!-- ============ APP ============ -->
   @if (showApp()) {
