+++ libs/adrian-badilla/frontend/landing-page/src/lib/components/reviews-section/reviews-section.component.scss
@@ -1,0 +1,74 @@
+// Testimonials carousel — local styles. Brand custom properties (--accent) are
+// inherited from the `.ab-root` ancestor rendered by the landing shell; the
+// marquee animation uses the global `ab-marqueeR` keyframes.
+:host {
+  display: block;
+}
+
+.rev-row:hover .rev-track { animation-play-state: paused; }
+
+.rev-card {
+  flex: none;
+  width: 370px;
+  display: flex;
+  flex-direction: column;
+  padding: 30px 30px 26px;
+  border-radius: 12px;
+  background: linear-gradient(180deg, rgba(255, 255, 255, 0.07), rgba(255, 255, 255, 0.02));
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  box-shadow: 0 24px 50px -28px rgba(0, 0, 0, 0.8);
+  transition: transform 0.35s ease, border-color 0.35s ease, box-shadow 0.35s ease;
+  &:hover {
+    transform: translateY(-6px);
+    border-color: hsl(356 88% 53% / 0.45);
+    box-shadow: 0 30px 60px -28px rgba(0, 0, 0, 0.9), 0 0 0 1px hsl(356 88% 53% / 0.25);
+  }
+  &__top { display: flex; align-items: center; justify-content: space-between; margin-bottom: 20px; }
+}
+
+.rev-g {
+  width: 26px;
+  height: 26px;
+  border-radius: 50%;
+  background: #fff;
+  color: #1a1a1a;
+  font-family: 'Anton', sans-serif;
+  font-size: 14px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.rev-quote {
+  font-size: 1.05rem;
+  line-height: 1.6;
+  color: rgba(255, 255, 255, 0.9);
+  margin: 0 0 26px;
+  flex: 1;
+  white-space: normal;
+  overflow-wrap: break-word;
+  &::before { content: '\201C'; color: hsl(356 88% 53% / 0.5); }
+  &::after { content: '\201D'; color: hsl(356 88% 53% / 0.5); }
+}
+
+.rev-foot { display: flex; align-items: center; gap: 14px; padding-top: 18px; border-top: 1px solid rgba(255, 255, 255, 0.1); }
+
+.rev-avatar {
+  flex: none;
+  width: 46px;
+  height: 46px;
+  border-radius: 50%;
+  background: hsl(356 88% 53% / 0.16);
+  color: var(--accent, hsl(356 88% 53%));
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-family: 'Anton', sans-serif;
+  font-size: 16px;
+}
+
+/* mobile: keep a single tidy row at a comfortable card width */
+@media (max-width: 820px) {
+  .rev-row + .rev-row { display: none !important; }
+  .rev-card { width: min(82vw, 340px); }
+}
